localization, venue update and a pro upgrade (#1156)
* Setting the stage for pro string updates throughout the Tragic Kingdom album

There are missing songs from the Tragic Kingdom album but this was driven by the PS3 visual enhancement with the new option

* localization, venue update and a pro update

AJR song "Bang" has already been merged into the main rb4_plus repository, but it hasn't been added here, so that's what was added.
Finalized updates to the venues for the Tragic Kingdom album.
Missing texts have been translated and improvements have been made to the writing.

* Add files via upload

* Delete _ark/songs/songs_updates.dta

* Add files via upload

* Add files via upload